refactor api auth and conditionalize load auth methods
