--- api/src/routes/auth.js
@@ -1,254 +1,0 @@
-const express = require('express');
-const { query, body } = require('express-validator');
-const IULoginHelper = require('@iusca/iulogin-helper');
-const config = require('config');
-const createError = require('http-errors');
-const { v4: uuidv4 } = require('uuid');
-const axios = require('axios');
-
-const { validate } = require('../middleware/validators');
-const asyncHandler = require('../middleware/asyncHandler');
-const { authenticate } = require('../middleware/auth');
-const { accessControl } = require('../middleware/auth');
-
-const userService = require('../services/user');
-const authService = require('../services/auth');
-const utils = require('../utils');
-
-const isPermittedTo = accessControl('auth');
-const router = express.Router();
-
-const IULogin = new IULoginHelper({
-  protocol: 'CAS',
-  mode: config.get('auth.mode'),
-});
-
-router.get(
-  '/cas/url',
-  validate([
-    query('service').notEmpty(),
-  ]),
-  asyncHandler(async (req, res, next) => {
-    // #swagger.tags = ['Auth']
-    const loginUrl = IULogin.get_login_url(req.query.service);
-    res.json({ url: loginUrl });
-  }),
-);
-
-router.post(
-  '/cas/verify',
-  validate([
-    body('ticket').notEmpty(),
-    body('service').notEmpty(),
-  ]),
-  (req, res, next) => {
-    // #swagger.tags = ['Auth']
-    // eslint-disable-next-line no-unused-vars
-    IULogin.validate(req.body.ticket, req.body.service, false, async (err, cas_id, profile) => {
-      if (err) return next(err);
-      try {
-        const user = await userService.findActiveUserBy('cas_id', cas_id);
-        if (user) {
-          const resObj = await authService.onLogin({ user });
-          return res.json(resObj);
-        }
-        // User was authenticated with CAS but they are not a portal user
-        // Send an empty success message
-        return res.status(204).send();
-      } catch (err2) {
-        return next(err2);
-      }
-    });
-  },
-);
-
-router.post('/refresh_token', authenticate, asyncHandler(async (req, res, next) => {
-  // #swagger.tags = ['Auth']
-  const user = await userService.findActiveUserBy('username', req.user.username);
-  if (user) {
-    const resObj = await authService.onLogin({ user });
-    return res.json(resObj);
-  }
-  // User has a valid token but they are deleted soon after and are not a portal user
-  // Send an invalid request error
-  return createError.BadRequest('Not a valid user');
-}));
-
-if (!['production', 'test'].includes(config.get('mode'))) {
-  router.post(
-    '/test_login',
-    asyncHandler(async (req, res, next) => {
-      // #swagger.tags = ['Auth']
-      if (req.body?.username === 'test_user') {
-        const user = await userService.findActiveUserBy('username', 'test_user');
-        const resObj = await authService.onLogin({ user });
-        return res.json(resObj);
-      }
-      return next(createError.Forbidden());
-    }),
-  );
-}
-
-router.post(
-  '/spoof/:username',
-  authenticate,
-  isPermittedTo('create'),
-  asyncHandler(async (req, res, next) => {
-  // #swagger.tags = ['Auth']
-    const user = await userService.findActiveUserBy('username', req.params.username);
-    const resObj = await authService.onLogin({ user, updateLastLogin: false });
-    return res.json(resObj);
-  }),
-);
-
-router.get(
-  '/google/url',
-  validate([
-    query('redirect_uri').notEmpty(),
-  ]),
-  asyncHandler(async (req, res, next) => {
-  // #swagger.tags = ['Auth']
-    const url = new URL(config.get('auth.google.authorization_endpoint'));
-
-    // https://developers.google.com/identity/openid-connect/openid-connect#state-param
-    const state = uuidv4();
-
-    const queryParams = {
-      response_type: 'code',
-      client_id: config.get('auth.google.client_id'),
-      redirect_uri: req.query.redirect_uri,
-      scope: config.get('auth.google.scope'),
-      state,
-    };
-
-    Object.keys(queryParams).forEach((key) => {
-      url.searchParams.set(key, queryParams[key]);
-    });
-
-    res.json({ url: url.toString(), state });
-  }),
-);
-
-// possible response codes:
-// 200 - return JWT
-// 204 - user authenticated but not a portal user
-// 500 - error
-router.post(
-  '/google/verify',
-  validate([
-    body('code').notEmpty(),
-    body('redirect_uri').notEmpty(),
-  ]),
-  asyncHandler(async (req, res, next) => {
-    // #swagger.tags = ['Auth']
-    // exchange code to get id token
-    const _res = await axios.post(
-      config.get('auth.google.token_endpoint'),
-      {
-        code: req.body.code,
-        client_id: config.get('auth.google.client_id'),
-        client_secret: config.get('auth.google.client_secret'),
-        redirect_uri: req.body.redirect_uri,
-        grant_type: 'authorization_code',
-      },
-    );
-    // decoded id_token looks like this:
-    // {
-    //   "iss": "accounts.google.com",
-    //   "sub": "100338926102874330519",
-    //   "email": "ddeepak6992@gmail.com",
-    //   "email_verified": true,
-    //   "at_hash": "soIBpOyDtjm3yw9633IB_A",
-    //   "iat": 1698353577,
-    //   "exp": 1698357177
-    // }
-    const id_data = utils.decodeJWT(_res.data.id_token);
-    const { email } = id_data;
-
-    const user = await userService.findActiveUserBy('email', email);
-    if (user) {
-      const resObj = await authService.onLogin({ user, method: 'Google' });
-      return res.json(resObj);
-    }
-    // User was authenticated with google but they are not a portal user
-    // Send an empty success message
-    return res.status(204).send();
-  }),
-);
-
-router.get(
-  '/cilogon/url',
-  validate([
-    query('redirect_uri').notEmpty(),
-  ]),
-  asyncHandler(async (req, res, next) => {
-  // #swagger.tags = ['Auth']
-    const url = new URL(config.get('auth.cilogon.authorization_endpoint'));
-
-    // https://developers.google.com/identity/openid-connect/openid-connect#state-param
-    const state = uuidv4();
-
-    const queryParams = {
-      response_type: 'code',
-      client_id: config.get('auth.cilogon.client_id'),
-      redirect_uri: req.query.redirect_uri,
-      scope: config.get('auth.cilogon.scope'),
-      state,
-    };
-
-    Object.keys(queryParams).forEach((key) => {
-      url.searchParams.set(key, queryParams[key]);
-    });
-
-    res.json({ url: url.toString(), state });
-  }),
-);
-
-// possible response codes:
-// 200 - return JWT
-// 204 - user authenticated but not a portal user
-// 500 - error
-router.post(
-  '/cilogon/verify',
-  validate([
-    body('code').notEmpty(),
-    body('redirect_uri').notEmpty(),
-  ]),
-  asyncHandler(async (req, res, next) => {
-    // #swagger.tags = ['Auth']
-    // exchange code to get id token
-    // POST with application/x-www-form-urlencoded data
-    const params = new URLSearchParams();
-    params.append('grant_type', 'authorization_code');
-    params.append('client_id', config.get('auth.cilogon.client_id'));
-    params.append('client_secret', config.get('auth.cilogon.client_secret'));
-    params.append('redirect_uri', req.body.redirect_uri);
-    params.append('code', req.body.code);
-    const _res = await axios.post(config.get('auth.cilogon.token_endpoint'), params);
-    // decoded id_token looks like this:
-    // {
-    //   iss: 'https://cilogon.org',
-    //   aud: 'cilogon:/client_id/',
-    //   exp: 1698419542,
-    //   iat: 1698418642,
-    //   jti: 'https://cilogon.org/oauth2/idToken/',
-    //   auth_time: 1698418641,
-    //   azp: 'cilogon:/client_id/',
-    //   sub: 'http://cilogon.org/serverE/users/',
-    //   acr: 'urn:oasis:names:tc:SAML:2.0:ac:classes:PasswordProtectedTransport',
-    //   email: 'deduggi@iu.edu'
-    // }
-    const id_data = utils.decodeJWT(_res.data.id_token);
-    const { email } = id_data;
-
-    const user = await userService.findActiveUserBy('email', email);
-    if (user) {
-      const resObj = await authService.onLogin({ user, method: 'CILogon' });
-      return res.json(resObj);
-    }
-    // User was authenticated with google but they are not a portal user
-    // Send an empty success message
-    return res.status(204).send();
-  }),
-);
-module.exports = router;
